Will raise exception when play twice onto a field
Thinking of the first thing to test is tough. This struck me as a good
place to start, as it tests the messages being sent to a user and
encourages an outside in testing approach.

The unit tests in game_spec meet the expectation in ui_spec, which
contains the feature being expected.

This change gives the Game class an attribute of cell/field, which is a
precursor to a board.