Game is instantiated with 'player' instances and a board
The rules of TTT demand two players. These have been formed into
separate Human and Computer classes that are passed into the Game class
on instantiation.

The Board and Players are passed as a series of hashes because it
removes the dependency on order of argument.

The Board is mocked and spied on within the Game test for assigning a
token to board to check that assignment methods are called, but the main
checking of assignment is done in the Board spec.

Furthermore, the mocking of the '[]=' method following a mock of 'state' on
the double<Board>, creates a warning about mocking on nil, which at this
point doesn't raise concern as I just want to check that the messages
are sent.