Let the Board know if a game is won by Row
This is a partial solution and first step; and minimal step towards
solving for diagonals and columns.

The Board::game_over? method slices it's board array into three
different arrays and then checks if any have three 'X' tokens in a row.

The tests are brittle because they only test for 'X'; they need to be
changed to allow any token that a player might have