UI can render the Board to the command line
The game needed an interface to allow players to choose a square. This
has been implemented within the UI class and run through the main.rb
file.

The main.rb file is starting to overreach its duties, so a good next
commit might involve refactoring the code within it into the
GameConstructor class