Refactor UI class
Remove test for type of game- it is not close to implementing and so is
unnecessary.
Refactor printing of the board to reduce method size and improve
readability. Note that horizontal dividers are calculated using the
formula for the nth term of a sequence.
Small change to main.rb to remove dead code and display more cleanly.